+++ src/app/globalCss.tsx
@@ -6,6 +6,8 @@ import { colorsToCss, themeToCss } from '../util';
 interface IGlobalCssProps {
   theme: ITheme;
   resetCss: string;
+  extraCss?: string;
+  isFullPageApp?: boolean;
 }
 
 export const GlobalCss = createGlobalStyle<IGlobalCssProps>`
@@ -28,6 +30,19 @@ export const GlobalCss = createGlobalStyle<IGlobalCssProps>`
     ${(props: IGlobalCssProps): string => themeToCss(props.theme.texts.default)};
     text-align: left;
   }
+
+  ${(props: IGlobalCssProps): string => (props.isFullPageApp ? `
+    html, body {
+      width: 100%;
+      height: 100%;
+      overscroll-behavior: none;
+    }
+
+    #root {
+      width: 100%;
+      height: 100%;
+    }
+  ` : '')};
 
   /* NOTE(krishan711): the :not(.button) needs to be specified as Buttons can act as links and these styles will be used on hover */
   /* since this overall ":hover" is more specific than the generic styles for the default button (with no modifier) */
@@ -104,4 +119,6 @@ export const GlobalCss = createGlobalStyle<IGlobalCssProps>`
       ${(props: IGlobalCssProps): string => themeToCss(props.theme.bulletTexts.default.normal.default.bullet)};
     }
   }
+
+  ${(props: IGlobalCssProps): string => props.extraCss || ''};
 `;